+++ cams-20200606/alibabacloud_cams20200606/models.py
@@ -20590,9 +20590,7 @@ class SendChatappMassMessageRequestSenderListFlowAction(TeaModel):
         flow_action_data: Dict[str, Any] = None,
         flow_token: str = None,
     ):
-        # The default parameter of the Flow.
         self.flow_action_data = flow_action_data
-        # The information about the Flow token.
         self.flow_token = flow_token
 
     def validate(self):
@@ -20624,7 +20622,6 @@ class SendChatappMassMessageRequestSenderListProductActionSectionsProductItems(T
         self,
         product_retailer_id: str = None,
     ):
-        # The retailer ID of the product.
         self.product_retailer_id = product_retailer_id
 
     def validate(self):
@@ -20653,9 +20650,7 @@ class SendChatappMassMessageRequestSenderListProductActionSections(TeaModel):
         product_items: List[SendChatappMassMessageRequestSenderListProductActionSectionsProductItems] = None,
         title: str = None,
     ):
-        # The products.
         self.product_items = product_items
-        # The name of the category.
         self.title = title
 
     def validate(self):
@@ -20696,9 +20691,7 @@ class SendChatappMassMessageRequestSenderListProductAction(TeaModel):
         sections: List[SendChatappMassMessageRequestSenderListProductActionSections] = None,
         thumbnail_product_retailer_id: str = None,
     ):
-        # The products. Up to 30 products and 10 categories can be added.
         self.sections = sections
-        # The retailer ID of the product.
         self.thumbnail_product_retailer_id = thumbnail_product_retailer_id
 
     def validate(self):
@@ -20742,17 +20735,10 @@ class SendChatappMassMessageRequestSenderList(TeaModel):
         template_params: Dict[str, str] = None,
         to: str = None,
     ):
-        # The Flow action.
         self.flow_action = flow_action
-        # The payload of the button.
         self.payload = payload
-        # The information about the product.
         self.product_action = product_action
-        # The parameters of the template.
         self.template_params = template_params
-        # The mobile phone number of the message receiver.
-        # 
-        # This parameter is required.
         self.to = to
 
     def validate(self):
@@ -20810,6 +20796,9 @@ class SendChatappMassMessageRequest(TeaModel):
         isv_code: str = None,
         label: str = None,
         language: str = None,
+        owner_id: int = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
         sender_list: List[SendChatappMassMessageRequestSenderList] = None,
         tag: str = None,
         task_id: str = None,
@@ -20817,59 +20806,28 @@ class SendChatappMassMessageRequest(TeaModel):
         template_name: str = None,
         ttl: int = None,
     ):
-        # The type of the channel. Valid values:
-        # 
-        # *   **whatsapp**\
-        # *   **viber**\
-        # *   **line** (under development)
-        # 
         # This parameter is required.
         self.channel_type = channel_type
-        # The space ID of the user.
         self.cust_space_id = cust_space_id
-        # The WhatsApp Business Account (WABA) ID of the RAM user within the independent software vendor (ISV) account.
-        # 
-        # >  CustWabaId is an obsolete parameter. Use CustSpaceId instead.
         self.cust_waba_id = cust_waba_id
-        # The content of the fallback message.
         self.fall_back_content = fall_back_content
-        # Specifies the period of time after which the fallback message is sent if the message receipt that indicates the message is delivered to clients is not received. If this parameter is left empty, the fallback message is sent only when the message fails to be sent or the message receipt that indicates the message is not delivered to clients is received. Unit: seconds. Valid values: 60 to 43200.
         self.fall_back_duration = fall_back_duration
-        # The ID of the fallback policy.
         self.fall_back_id = fall_back_id
-        # The fallback rule. Valid values:
-        # 
-        # *   **undelivered**: A fallback is triggered if the message is not delivered to clients. When the message is being sent, the template parameters are verified. If the parameters fail to pass the verification, the message fails to be sent. Whether the template and phone number are prohibited is not verified. By default, this value is used when FallBackRule is left empty.
-        # *   **sentFailed**: A fallback is triggered even if the template parameters including variables fail to pass the verification. If the channelType, type, messageType, to, and from parameters fail to pass the verification, a fallback is not triggered.
         self.fall_back_rule = fall_back_rule
-        # The mobile phone number of the message sender.
-        # 
         # This parameter is required.
         self.from_ = from_
-        # The ISV verification code. This parameter is used to verify whether the RAM user is authorized by the ISV account.
         self.isv_code = isv_code
-        # The type of the Viber message. Valid values:
-        # 
-        # *   **promotion**\
-        # *   **transaction**\
         self.label = label
-        # The language. For more information about language codes, see [Language codes](https://help.aliyun.com/document_detail/463420.html).
-        # 
         # This parameter is required.
         self.language = language
-        # The mobile phone numbers of the message receivers.
-        # 
-        # This parameter is required.
+        self.owner_id = owner_id
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
         self.sender_list = sender_list
-        # The tag information when the ChannelType parameter is set to viber.
         self.tag = tag
-        # The task ID.
         self.task_id = task_id
-        # The template code.
         self.template_code = template_code
-        # The name of the message template.
         self.template_name = template_name
-        # The timeout period for sending messages when the ChannelType parameter is set to viber. Valid values: 30 to 1209600. Unit: seconds.
         self.ttl = ttl
 
     def validate(self):
@@ -20906,6 +20864,12 @@ class SendChatappMassMessageRequest(TeaModel):
             result['Label'] = self.label
         if self.language is not None:
             result['Language'] = self.language
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
         result['SenderList'] = []
         if self.sender_list is not None:
             for k in self.sender_list:
@@ -20946,6 +20910,12 @@ class SendChatappMassMessageRequest(TeaModel):
             self.label = m.get('Label')
         if m.get('Language') is not None:
             self.language = m.get('Language')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
         self.sender_list = []
         if m.get('SenderList') is not None:
             for k in m.get('SenderList'):
@@ -20978,6 +20948,9 @@ class SendChatappMassMessageShrinkRequest(TeaModel):
         isv_code: str = None,
         label: str = None,
         language: str = None,
+        owner_id: int = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
         sender_list_shrink: str = None,
         tag: str = None,
         task_id: str = None,
@@ -20985,59 +20958,28 @@ class SendChatappMassMessageShrinkRequest(TeaModel):
         template_name: str = None,
         ttl: int = None,
     ):
-        # The type of the channel. Valid values:
-        # 
-        # *   **whatsapp**\
-        # *   **viber**\
-        # *   **line** (under development)
-        # 
         # This parameter is required.
         self.channel_type = channel_type
-        # The space ID of the user.
         self.cust_space_id = cust_space_id
-        # The WhatsApp Business Account (WABA) ID of the RAM user within the independent software vendor (ISV) account.
-        # 
-        # >  CustWabaId is an obsolete parameter. Use CustSpaceId instead.
         self.cust_waba_id = cust_waba_id
-        # The content of the fallback message.
         self.fall_back_content = fall_back_content
-        # Specifies the period of time after which the fallback message is sent if the message receipt that indicates the message is delivered to clients is not received. If this parameter is left empty, the fallback message is sent only when the message fails to be sent or the message receipt that indicates the message is not delivered to clients is received. Unit: seconds. Valid values: 60 to 43200.
         self.fall_back_duration = fall_back_duration
-        # The ID of the fallback policy.
         self.fall_back_id = fall_back_id
-        # The fallback rule. Valid values:
-        # 
-        # *   **undelivered**: A fallback is triggered if the message is not delivered to clients. When the message is being sent, the template parameters are verified. If the parameters fail to pass the verification, the message fails to be sent. Whether the template and phone number are prohibited is not verified. By default, this value is used when FallBackRule is left empty.
-        # *   **sentFailed**: A fallback is triggered even if the template parameters including variables fail to pass the verification. If the channelType, type, messageType, to, and from parameters fail to pass the verification, a fallback is not triggered.
         self.fall_back_rule = fall_back_rule
-        # The mobile phone number of the message sender.
-        # 
         # This parameter is required.
         self.from_ = from_
-        # The ISV verification code. This parameter is used to verify whether the RAM user is authorized by the ISV account.
         self.isv_code = isv_code
-        # The type of the Viber message. Valid values:
-        # 
-        # *   **promotion**\
-        # *   **transaction**\
         self.label = label
-        # The language. For more information about language codes, see [Language codes](https://help.aliyun.com/document_detail/463420.html).
-        # 
         # This parameter is required.
         self.language = language
-        # The mobile phone numbers of the message receivers.
-        # 
-        # This parameter is required.
+        self.owner_id = owner_id
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
         self.sender_list_shrink = sender_list_shrink
-        # The tag information when the ChannelType parameter is set to viber.
         self.tag = tag
-        # The task ID.
         self.task_id = task_id
-        # The template code.
         self.template_code = template_code
-        # The name of the message template.
         self.template_name = template_name
-        # The timeout period for sending messages when the ChannelType parameter is set to viber. Valid values: 30 to 1209600. Unit: seconds.
         self.ttl = ttl
 
     def validate(self):
@@ -21071,6 +21013,12 @@ class SendChatappMassMessageShrinkRequest(TeaModel):
             result['Label'] = self.label
         if self.language is not None:
             result['Language'] = self.language
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
         if self.sender_list_shrink is not None:
             result['SenderList'] = self.sender_list_shrink
         if self.tag is not None:
@@ -21109,6 +21057,12 @@ class SendChatappMassMessageShrinkRequest(TeaModel):
             self.label = m.get('Label')
         if m.get('Language') is not None:
             self.language = m.get('Language')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
         if m.get('SenderList') is not None:
             self.sender_list_shrink = m.get('SenderList')
         if m.get('Tag') is not None:
@@ -21230,9 +21184,7 @@ class SendChatappMessageRequestFlowAction(TeaModel):
         flow_action_data: Dict[str, Any] = None,
         flow_token: str = None,
     ):
-        # The default parameter of the Flow.
         self.flow_action_data = flow_action_data
-        # The Flow token.
         self.flow_token = flow_token
 
     def validate(self):
@@ -21264,7 +21216,6 @@ class SendChatappMessageRequestProductActionSectionsProductItems(TeaModel):
         self,
         product_retailer_id: str = None,
     ):
-        # The retailer ID of the product.
         self.product_retailer_id = product_retailer_id
 
     def validate(self):
@@ -21293,9 +21244,7 @@ class SendChatappMessageRequestProductActionSections(TeaModel):
         product_items: List[SendChatappMessageRequestProductActionSectionsProductItems] = None,
         title: str = None,
     ):
-        # The products.
         self.product_items = product_items
-        # The name of the category.
         self.title = title
 
     def validate(self):
@@ -21336,9 +21285,7 @@ class SendChatappMessageRequestProductAction(TeaModel):
         sections: List[SendChatappMessageRequestProductActionSections] = None,
         thumbnail_product_retailer_id: str = None,
     ):
-        # The products. Up to 30 products and 10 categories can be added.
         self.sections = sections
-        # The retailer ID of the product.
         self.thumbnail_product_retailer_id = thumbnail_product_retailer_id
 
     def validate(self):
@@ -21391,9 +21338,12 @@ class SendChatappMessageRequest(TeaModel):
         label: str = None,
         language: str = None,
         message_type: str = None,
+        owner_id: int = None,
         payload: List[str] = None,
         product_action: SendChatappMessageRequestProductAction = None,
         recipient_type: str = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
         tag: str = None,
         task_id: str = None,
         template_code: str = None,
@@ -21404,12 +21354,6 @@ class SendChatappMessageRequest(TeaModel):
         ttl: int = None,
         type: str = None,
     ):
-        # The channel type. Valid values:
-        # 
-        # *   **whatsapp**\
-        # *   **viber**\
-        # *   **line** (under development)
-        # 
         # This parameter is required.
         self.channel_type = channel_type
         # The message content.
@@ -21438,99 +21382,36 @@ class SendChatappMessageRequest(TeaModel):
         # *   If you set **messageType** to **text_video**, you must specify **text**, **link**, **thumbnail**, **fileSize**, and **duration**.
         # *   If you set **messageType** to **text_video_button**, you must specify **text**, **link**, **thumbnail**, **fileSize**, **duration**, and **caption**. In addition, you must not specify **action**.
         self.content = content
-        # The ID of the reply message.
         self.context_message_id = context_message_id
-        # The space ID of the user.
         self.cust_space_id = cust_space_id
-        # The WhatsApp Business Account (WABA) ID of the RAM user within the independent software vendor (ISV) account.
-        # 
-        # >  CustWabaId is an obsolete parameter. Use CustSpaceId instead.
         self.cust_waba_id = cust_waba_id
-        # The content of the fallback message.
         self.fall_back_content = fall_back_content
-        # Specifies the period of time after which the fallback message is sent if the message receipt that indicates the message is delivered to clients is not received. If this parameter is left empty, the fallback message is sent only when the **message fails to be sent** or **the message receipt that indicates the message is not delivered to clients** is received. Unit: seconds. Valid values: 60 to 43200.
         self.fall_back_duration = fall_back_duration
-        # The ID of the fallback policy. You can create a fallback policy and view the information in the Chat App Message Service console.
         self.fall_back_id = fall_back_id
-        # The fallback rule. Valid values:
-        # 
-        # *   **undelivered**: A fallback is triggered if the message is not delivered to clients. When the message is being sent, the template parameters are verified. If the parameters fail to pass the verification, the message fails to be sent. Whether the template and phone number are prohibited is not verified. By default, this value is used when FallBackRule is left empty.
-        # *   **sentFailed**: A fallback is triggered even if the template parameters including variables fail to pass the verification. If the channelType, type, messageType, to, and from parameters fail to pass the verification, a fallback is not triggered.
         self.fall_back_rule = fall_back_rule
-        # The Flow action.
         self.flow_action = flow_action
-        # The mobile phone number of the message sender.
-        # 
-        # >  You can specify a mobile phone number that is registered for a WhatsApp account and is approved in the Chat App Message Service console.
-        # 
         # This parameter is required.
         self.from_ = from_
-        # The ISV verification code. This parameter is used to verify whether the RAM user is authorized by the ISV account.
         self.isv_code = isv_code
-        # The type of the Viber message. This parameter is required if ChannelType is set to viber. Valid values:
-        # 
-        # *   **promotion**\
-        # *   **transaction**\
         self.label = label
-        # The language that is used in the message template. This parameter is required only if you set the Type parameter to **template**. For more information about language codes, see [Language codes](https://help.aliyun.com/document_detail/463420.html).
         self.language = language
-        # The specific type of the message. This parameter is required only if you set the Type parameter to **message**.
-        # 
-        # **Valid values of MessageType when you set the ChannelType parameter to whatsapp:**\
-        # 
-        # *   **text**: a text message.
-        # *   **image**: an image message.
-        # *   **video**: a video message.
-        # *   **audio**: an audio message.
-        # *   **document**: a document message.
-        # *   **interactive**: an interactive message.
-        # *   **contacts**: a contact message.
-        # *   **location**: a location message.
-        # *   **sticker**: a sticker message.
-        # *   **reaction**: a reaction message.
-        # 
-        # **Valid values of MessageType when you set the ChannelType parameter to viber:**\
-        # 
-        # *   **text**: a text message.
-        # *   **image**: an image message.
-        # *   **video**: a video message.
-        # *   **document**: a document message.
-        # *   **text_button**: a message that contains the text and button media objects.
-        # *   **text_image_button**: a message that contains multiple media objects, including the text, image, and button.
-        # *   **text_video**: a message that contains the text and video media objects.
-        # *   **text_video_button**: a message that contains multiple media objects, including text, video, and button.
-        # *   **text_image**: a message that contains the text and image media objects.
-        # 
-        # > For more information, see [Parameters of a message template](https://help.aliyun.com/document_detail/454530.html).
         self.message_type = message_type
+        self.owner_id = owner_id
         # The payload of the button.
         self.payload = payload
-        # The information about the products included in the WhatsApp catalog message or multi-product message (MPM).
         self.product_action = product_action
         self.recipient_type = recipient_type
-        # The tag information of the Viber message.
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
         self.tag = tag
-        # The task ID.
         self.task_id = task_id
-        # The code of the message template. This parameter is required only if you set the Type parameter to **template**.
         self.template_code = template_code
-        # The name of the message template.
         self.template_name = template_name
-        # The variables of the message template.
         self.template_params = template_params
-        # The mobile phone number of the message receiver.
-        # 
         # This parameter is required.
         self.to = to
-        # The tracking data of the Viber message.
         self.tracking_data = tracking_data
-        # The timeout period for sending the Viber message. Valid values: 30 to 1209600. Unit: seconds.
         self.ttl = ttl
-        # The message type. Valid values:
-        # 
-        # *   **template**: the template message. A template message is sent based on a template that is created and approved in the Chat App Message Service console. You can send template messages based on your business requirements.
-        # *   **message**: the custom message. You can send a custom WhatsApp message to a user only within 24 hours after you receive the last message from the user. This limit does not apply to custom Viber messages.
-        # 
         # This parameter is required.
         self.type = type
 
@@ -21576,12 +21457,18 @@ class SendChatappMessageRequest(TeaModel):
             result['Language'] = self.language
         if self.message_type is not None:
             result['MessageType'] = self.message_type
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
         if self.payload is not None:
             result['Payload'] = self.payload
         if self.product_action is not None:
             result['ProductAction'] = self.product_action.to_map()
         if self.recipient_type is not None:
             result['RecipientType'] = self.recipient_type
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
         if self.tag is not None:
             result['Tag'] = self.tag
         if self.task_id is not None:
@@ -21635,6 +21522,8 @@ class SendChatappMessageRequest(TeaModel):
             self.language = m.get('Language')
         if m.get('MessageType') is not None:
             self.message_type = m.get('MessageType')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
         if m.get('Payload') is not None:
             self.payload = m.get('Payload')
         if m.get('ProductAction') is not None:
@@ -21642,6 +21531,10 @@ class SendChatappMessageRequest(TeaModel):
             self.product_action = temp_model.from_map(m['ProductAction'])
         if m.get('RecipientType') is not None:
             self.recipient_type = m.get('RecipientType')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
         if m.get('Tag') is not None:
             self.tag = m.get('Tag')
         if m.get('TaskId') is not None:
@@ -21681,9 +21574,12 @@ class SendChatappMessageShrinkRequest(TeaModel):
         label: str = None,
         language: str = None,
         message_type: str = None,
+        owner_id: int = None,
         payload_shrink: str = None,
         product_action_shrink: str = None,
         recipient_type: str = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
         tag: str = None,
         task_id: str = None,
         template_code: str = None,
@@ -21694,12 +21590,6 @@ class SendChatappMessageShrinkRequest(TeaModel):
         ttl: int = None,
         type: str = None,
     ):
-        # The channel type. Valid values:
-        # 
-        # *   **whatsapp**\
-        # *   **viber**\
-        # *   **line** (under development)
-        # 
         # This parameter is required.
         self.channel_type = channel_type
         # The message content.
@@ -21728,99 +21618,36 @@ class SendChatappMessageShrinkRequest(TeaModel):
         # *   If you set **messageType** to **text_video**, you must specify **text**, **link**, **thumbnail**, **fileSize**, and **duration**.
         # *   If you set **messageType** to **text_video_button**, you must specify **text**, **link**, **thumbnail**, **fileSize**, **duration**, and **caption**. In addition, you must not specify **action**.
         self.content = content
-        # The ID of the reply message.
         self.context_message_id = context_message_id
-        # The space ID of the user.
         self.cust_space_id = cust_space_id
-        # The WhatsApp Business Account (WABA) ID of the RAM user within the independent software vendor (ISV) account.
-        # 
-        # >  CustWabaId is an obsolete parameter. Use CustSpaceId instead.
         self.cust_waba_id = cust_waba_id
-        # The content of the fallback message.
         self.fall_back_content = fall_back_content
-        # Specifies the period of time after which the fallback message is sent if the message receipt that indicates the message is delivered to clients is not received. If this parameter is left empty, the fallback message is sent only when the **message fails to be sent** or **the message receipt that indicates the message is not delivered to clients** is received. Unit: seconds. Valid values: 60 to 43200.
         self.fall_back_duration = fall_back_duration
-        # The ID of the fallback policy. You can create a fallback policy and view the information in the Chat App Message Service console.
         self.fall_back_id = fall_back_id
-        # The fallback rule. Valid values:
-        # 
-        # *   **undelivered**: A fallback is triggered if the message is not delivered to clients. When the message is being sent, the template parameters are verified. If the parameters fail to pass the verification, the message fails to be sent. Whether the template and phone number are prohibited is not verified. By default, this value is used when FallBackRule is left empty.
-        # *   **sentFailed**: A fallback is triggered even if the template parameters including variables fail to pass the verification. If the channelType, type, messageType, to, and from parameters fail to pass the verification, a fallback is not triggered.
         self.fall_back_rule = fall_back_rule
-        # The Flow action.
         self.flow_action_shrink = flow_action_shrink
-        # The mobile phone number of the message sender.
-        # 
-        # >  You can specify a mobile phone number that is registered for a WhatsApp account and is approved in the Chat App Message Service console.
-        # 
         # This parameter is required.
         self.from_ = from_
-        # The ISV verification code. This parameter is used to verify whether the RAM user is authorized by the ISV account.
         self.isv_code = isv_code
-        # The type of the Viber message. This parameter is required if ChannelType is set to viber. Valid values:
-        # 
-        # *   **promotion**\
-        # *   **transaction**\
         self.label = label
-        # The language that is used in the message template. This parameter is required only if you set the Type parameter to **template**. For more information about language codes, see [Language codes](https://help.aliyun.com/document_detail/463420.html).
         self.language = language
-        # The specific type of the message. This parameter is required only if you set the Type parameter to **message**.
-        # 
-        # **Valid values of MessageType when you set the ChannelType parameter to whatsapp:**\
-        # 
-        # *   **text**: a text message.
-        # *   **image**: an image message.
-        # *   **video**: a video message.
-        # *   **audio**: an audio message.
-        # *   **document**: a document message.
-        # *   **interactive**: an interactive message.
-        # *   **contacts**: a contact message.
-        # *   **location**: a location message.
-        # *   **sticker**: a sticker message.
-        # *   **reaction**: a reaction message.
-        # 
-        # **Valid values of MessageType when you set the ChannelType parameter to viber:**\
-        # 
-        # *   **text**: a text message.
-        # *   **image**: an image message.
-        # *   **video**: a video message.
-        # *   **document**: a document message.
-        # *   **text_button**: a message that contains the text and button media objects.
-        # *   **text_image_button**: a message that contains multiple media objects, including the text, image, and button.
-        # *   **text_video**: a message that contains the text and video media objects.
-        # *   **text_video_button**: a message that contains multiple media objects, including text, video, and button.
-        # *   **text_image**: a message that contains the text and image media objects.
-        # 
-        # > For more information, see [Parameters of a message template](https://help.aliyun.com/document_detail/454530.html).
         self.message_type = message_type
+        self.owner_id = owner_id
         # The payload of the button.
         self.payload_shrink = payload_shrink
-        # The information about the products included in the WhatsApp catalog message or multi-product message (MPM).
         self.product_action_shrink = product_action_shrink
         self.recipient_type = recipient_type
-        # The tag information of the Viber message.
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
         self.tag = tag
-        # The task ID.
         self.task_id = task_id
-        # The code of the message template. This parameter is required only if you set the Type parameter to **template**.
         self.template_code = template_code
-        # The name of the message template.
         self.template_name = template_name
-        # The variables of the message template.
         self.template_params_shrink = template_params_shrink
-        # The mobile phone number of the message receiver.
-        # 
         # This parameter is required.
         self.to = to
-        # The tracking data of the Viber message.
         self.tracking_data = tracking_data
-        # The timeout period for sending the Viber message. Valid values: 30 to 1209600. Unit: seconds.
         self.ttl = ttl
-        # The message type. Valid values:
-        # 
-        # *   **template**: the template message. A template message is sent based on a template that is created and approved in the Chat App Message Service console. You can send template messages based on your business requirements.
-        # *   **message**: the custom message. You can send a custom WhatsApp message to a user only within 24 hours after you receive the last message from the user. This limit does not apply to custom Viber messages.
-        # 
         # This parameter is required.
         self.type = type
 
@@ -21863,12 +21690,18 @@ class SendChatappMessageShrinkRequest(TeaModel):
             result['Language'] = self.language
         if self.message_type is not None:
             result['MessageType'] = self.message_type
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
         if self.payload_shrink is not None:
             result['Payload'] = self.payload_shrink
         if self.product_action_shrink is not None:
             result['ProductAction'] = self.product_action_shrink
         if self.recipient_type is not None:
             result['RecipientType'] = self.recipient_type
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
         if self.tag is not None:
             result['Tag'] = self.tag
         if self.task_id is not None:
@@ -21921,12 +21754,18 @@ class SendChatappMessageShrinkRequest(TeaModel):
             self.language = m.get('Language')
         if m.get('MessageType') is not None:
             self.message_type = m.get('MessageType')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
         if m.get('Payload') is not None:
             self.payload_shrink = m.get('Payload')
         if m.get('ProductAction') is not None:
             self.product_action_shrink = m.get('ProductAction')
         if m.get('RecipientType') is not None:
             self.recipient_type = m.get('RecipientType')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
         if m.get('Tag') is not None:
             self.tag = m.get('Tag')
         if m.get('TaskId') is not None:
@@ -21951,11 +21790,13 @@ class SendChatappMessageShrinkRequest(TeaModel):
 class SendChatappMessageResponseBody(TeaModel):
     def __init__(
         self,
+        access_denied_detail: str = None,
         code: str = None,
         message: str = None,
         message_id: str = None,
         request_id: str = None,
     ):
+        self.access_denied_detail = access_denied_detail
         # The HTTP status code returned.
         # 
         # *   A value of OK indicates that the call is successful.
@@ -21977,6 +21818,8 @@ class SendChatappMessageResponseBody(TeaModel):
             return _map
 
         result = dict()
+        if self.access_denied_detail is not None:
+            result['AccessDeniedDetail'] = self.access_denied_detail
         if self.code is not None:
             result['Code'] = self.code
         if self.message is not None:
@@ -21989,6 +21832,8 @@ class SendChatappMessageResponseBody(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AccessDeniedDetail') is not None:
+            self.access_denied_detail = m.get('AccessDeniedDetail')
         if m.get('Code') is not None:
             self.code = m.get('Code')
         if m.get('Message') is not None: